refactor: 대기방 훅 책임 분리 및 controller 레이어 분리
